added version to flow data

+++ includes/Services/FlowService.php
@@ -21,8 +21,10 @@ class FlowService {
 		if ( ! ( $flow_data ) ) {
 			return \update_option( Options::get_option_name( 'flow' ), $default_flow_data );
 		}
-		$updated_flow_data = self::update_flow_data_recursive( $default_flow_data, $flow_data );
-		return \update_option( Options::get_option_name( 'flow' ), $updated_flow_data );
+		if ( ! array_key_exists( 'version' , $flow_data) || strcmp( $flow_data['version'] , $default_flow_data['version'] ) !== 0 ) {
+			$updated_flow_data = self::update_flow_data_recursive( $default_flow_data, $flow_data );
+			return \update_option( Options::get_option_name( 'flow' ), $updated_flow_data );
+		}
 	}
 
 	/**
@@ -68,6 +70,12 @@ class FlowService {
 		$exception_list    = Flows::get_exception_list();
 		$updated_flow_data = array();
 		foreach ( $default_flow_data as $key => $value ) {
+			// To update the options with the recent version of flow data
+			if( strcmp( 'version' , $key ) === 0 ) {
+				$updated_flow_data[ $key ] = $value;
+				continue;
+			}
+
 			// Any Key renamed is updated in the database with NewKey and the value from the OldKey is retained or not based on retain_existing_value
 			if ( count( $flow_data_fixes ) > 0 ) {
 				foreach ( $flow_data_fixes as $index => $fix ) {
